refactor: reorganize components and styles into product-card and draw directories
