+++ app/components/draw/PixelArtCanvas.tsx
